Moved EnemyCore into appropriate directory
