+++ Scenes/Enemies/Enemy.cs
@@ -3,6 +3,7 @@ using System;
 using System.Linq;
 using TheLoneLanternProject;
 using TheLoneLanternProject.Constants;
+using TheLoneLanternProject.Scenes.Enemies;
 
 public partial class Enemy : EnemyCore
 {
